Remove excessive private submod.
